iz-coach/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
openai==1.76.*
pydantic-settings==2.7.*
python-multipart==0.0.20
pypdf==5.4.*
python-docx==1.1.*
